requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
python-dotenv==1.0.0
python-telegram-bot==20.3
werkzeug==2.3.6